[TASK] Array for bullet list should be according to type
The controller is producing a multidimensional array for all types in content element "Bullet list". This is done for the added definition list. For <ul> and <ol> a strange construct is needed this way.

This patch changes the behaviour by sending a simple array to the view for <ol> and <ul> and a multidimensional array for <dl>, which has two columns.