[TASK] CE "File links" is using field "content element layout" for several displaying options
The content element "File links" is using this field for displaying only the file name, the file name with an extension icon, or with a thumbnail

The field "content element layout" is misused here. There should be a different field for this kind of display options. There is already a checkbox for displaying the file size.

This patch adds a new field for the choice "only filename", "Filename with extension icon" and "Filename with thumbnail" and adds an extra checkbox to enable the display of the description