[TASK] Get rid of the FlexForm in CE "Table"
No content element is using a FlexForm, except for "Table". These fields can easily move to the database.

Some fields can be deleted:
Summary: Not available anymore in HTML5
No styles: Can be done with CSS
CSS class: Not for an editor, and the layout field can be used

This patch removes the FlexForm and the mentioned fields