[TASK] Show minimum set of editing fields by default
Most beginners will start with an admin account. A lot of fields are visible for these editors, which are not used by default without taking any other action - like frontend user access, or are not often used.

This patch adds configuration to the Extension Manager where a minimum set of fields has been selected by default. The more educated users can turn this option off, so all the editable fields of content elements are available for administrators and regular be_groups, if exclude fields are allowed.