[TASK] Replace CE's image, text & textpic with textmedia
Frans Saris is working on a FAL online media connector. With this we can load YouTube and Vimeo movies into a content element. This is suitable for the gallery as well, so you can combine videos and images into one content element.

This patch replaces the content elements "Image", "Text" and "Text & Images" with the new "Text & Media" and includes the dependency on "fal_online_media_connector"